- #110: Move tech types to techs/ folder.
