+++ src/app/hanzhong/techs/TechLevelType.ts
@@ -1,0 +1,7 @@
+import type { HanzhongBonusType } from '../types';
+import type { HanzhongTechResourceType } from './TechResourceType';
+
+export type HanzhongTechLevelType = {
+  resources: HanzhongTechResourceType;
+  bonuses: HanzhongBonusType;
+};
