+++ src/app/hanzhong/techs/TechResourceType.ts
