+++ src/app/hanzhong/techs/info/Info.tsx
@@ -6,16 +6,16 @@ import DialogTitle from '@mui/material/DialogTitle';
 import Grid from '@mui/material/Grid';
 
 import { getTechById } from '../../data';
-import type { HanzhongTechType } from '../../types';
 import { NO_TECH_ID } from '../constants';
 import { useHanzhongTechsContext } from '../context';
 import { HanzhongTechsTechCardImage } from '../HanzhongTechsTechCardImage';
 import { useHanzhongContext } from '../../HanzhongContext';
 import { generateLevelsToDisplay } from './generate-levels-to-display';
-import type { HanzhongTechsTechLevelDisplayType } from './types';
 import { HanzhongTechsTechInfoLevel } from './level';
 import Typography from '@mui/material/Typography';
 import { HanzhongTechsTechInfoRequirements } from './InfoRequirements';
+import type { HanzhongTechType } from '../TechType';
+import type { HanzhongTechsTechLevelDisplayType } from '../TechsTechLevelDisplayType';
 
 export type HanzhongTechsTechInfoProps = {
   onClose: () => void;
