+++ src/app/hanzhong/techs/info/generate-levels-to-display.ts
@@ -1,8 +1,7 @@
-import type { HanzhongTechLevelType } from '../../types';
-
+import type { HanzhongTechLevelType } from '../TechLevelType';
+import type { HanzhongTechsTechLevelDisplayType } from '../TechsTechLevelDisplayType';
 import { calculateDeltaBonuses } from './calculate-delta-bonuses';
 import { sumTechResources } from './sum-tech-resources';
-import type { HanzhongTechsTechLevelDisplayType } from './types';
 
 export const generateLevelsToDisplay = (
   levels: HanzhongTechLevelType[],
