--- src/app/hanzhong/techs/info/types/index.ts
@@ -1,1 +1,0 @@
-export { type HanzhongTechsTechLevelDisplayType } from './display-type';
