--- src/app/hanzhong/types/hanzhong-tech-level-type.ts
@@ -1,7 +1,0 @@
-import type { HanzhongBonusType } from './hanzhong-bonus-type';
-import type { HanzhongTechResourceType } from './hanzhong-tech-resource-type';
-
-export type HanzhongTechLevelType = {
-  resources: HanzhongTechResourceType;
-  bonuses: HanzhongBonusType;
-};
