+++ src/app/hanzhong/utils/find-resources-needed-by-tech-id-and-level.ts
@@ -1,5 +1,5 @@
 import { getTechById } from '../data';
-import type { HanzhongTechResourceType } from '../types';
+import type { HanzhongTechResourceType } from '../techs';
 
 export const findResourcesNeededByTechIdAndLevel = (techId: string, level: number): HanzhongTechResourceType | -1 | null => {
   const tech = getTechById(techId);
